drastically simplify the wasmEnvironment and wasmInstance internals
