+++ rwasm/ffi_environment.go
@@ -1,12 +1,7 @@
 package rwasm
 
 import (
-	"context"
-	"crypto/rand"
-	"math"
-	"math/big"
 	"sync"
-	"time"
 
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
@@ -14,7 +9,6 @@ import (
 	"github.com/suborbital/reactr/rwasm/moduleref"
 	"github.com/suborbital/vektor/vlog"
 	"github.com/wasmerio/wasmer-go/wasmer"
-	"golang.org/x/sync/semaphore"
 )
 
 /*
@@ -22,90 +16,40 @@ import (
  an FFI API. Functions exported from a WASM module can be easily called by Go code via the Wasmer instance exports, but returning data
  to the host Go code is not quite as straightforward.
 
- In order to accomplish this, rwasm internally keeps a set of "environments" in a singleton package var (`environments` below).
- Each environment is a container that includes the WASM module bytes, and a set of WASM instances (runtimes) to execute said module.
- The envionment object has an index referencing its place in the singleton array, and each instance has an index referencing its position within
- the environment's instance array.
+ In order to accomplish this, rwasm creates 'wasmEnvironments' which represent a single instantiated module. Each environment contains a pool of
+ 'wasmInstances', which are provided for use on a rotating basis. Instances can be added and removed from the pool as needed by the `wasmRunner`.
 
- When a WASM function calls one of the FFI API functions, it includes the `ident`` value that was provided at the beginning
- of job execution, which allows rwasm to look up the [env][instance] and send the result on the appropriate result channel. This is needed due to
- the way Go makes functions available on the FFI using CGO.
+ When a WASM function calls one of the FFI API functions, it includes the `ident` value that was provided at the beginning
+ of job execution, which allows rwasm to look up the wasmInstance from the global `instanceMapper` and send the result on
+ the appropriate result channel. This is needed due to the way Go makes functions available to the FFI.
 */
-
-// the globally shared set of Wasm environments, accessed by UUID
-var environments = map[string]*wasmEnvironment{}
-
-// a lock to ensure the environments array is concurrency safe (didn't use sync.Map to prevent type coersion)
-var envLock = sync.RWMutex{}
-
-// the instance mapper maps a random int32 to a wasm instance to prevent malicious access to other instances via the FFI
-var instanceMapper = sync.Map{}
 
 // the internal Logger used by the Wasm runtime system
 var internalLogger = vlog.Default()
 
-// wasmEnvironment is an environmenr in which Wasm instances run
+// wasmEnvironment is an environment in which Wasm instances run
 type wasmEnvironment struct {
-	UUID      string
-	ref       *moduleref.WasmModuleRef
-	module    *wasmer.Module
-	store     *wasmer.Store
-	imports   *wasmer.ImportObject
-	instances []*wasmInstance
+	UUID    string
+	ref     *moduleref.WasmModuleRef
+	module  *wasmer.Module
+	store   *wasmer.Store
+	imports *wasmer.ImportObject
 
-	// the index of the last used wasm instance
-	instIndex int
-	lock      sync.RWMutex
+	availableInstances chan *wasmInstance
+
+	lock sync.RWMutex
 }
 
-type wasmInstance struct {
-	wasmerInst *wasmer.Instance
-
-	ctx *rt.Ctx
-
-	ffiResult []byte
-
-	resultChan chan []byte
-	errChan    chan rt.RunErr
-
-	access *semaphore.Weighted
-}
-
-// instanceReference is a "pointer" to the global environments array and the
-// wasm instances within each environment
-type instanceReference struct {
-	EnvUUID   string
-	InstIndex int
-}
-
-// newEnvironment creates a new environment and adds it to the shared environments array
-// such that Wasm instances can return data to the correct place
+// newEnvironment creates a new environment with a pool of available wasmInstances
 func newEnvironment(ref *moduleref.WasmModuleRef) *wasmEnvironment {
-	envLock.Lock()
-	defer envLock.Unlock()
-
 	e := &wasmEnvironment{
-		UUID:      uuid.New().String(),
-		ref:       ref,
-		instances: []*wasmInstance{},
-		instIndex: 0,
-		lock:      sync.RWMutex{},
+		UUID:               uuid.New().String(),
+		ref:                ref,
+		availableInstances: make(chan *wasmInstance, 32),
+		lock:               sync.RWMutex{},
 	}
-
-	environments[e.UUID] = e
 
 	return e
-}
-
-func (w *wasmEnvironment) instanceAtIndex(idx int) (*wasmInstance, error) {
-	w.lock.RLock()
-	defer w.lock.RUnlock()
-
-	if len(w.instances) <= idx-1 {
-		return nil, errors.New("invalid instance index")
-	}
-
-	return w.instances[idx], nil
 }
 
 // addInstance adds a new Wasm instance to the environment's pool
@@ -151,47 +95,17 @@ func (w *wasmEnvironment) addInstance() error {
 		wasmerInst: inst,
 		resultChan: make(chan []byte, 1),
 		errChan:    make(chan rt.RunErr, 1),
-		access:     semaphore.NewWeighted(1),
 	}
 
-	w.instances = append(w.instances, instance)
+	w.availableInstances <- instance
 
 	return nil
 }
 
 func (w *wasmEnvironment) removeInstance() error {
-	// this process happens in a 4 step lock-dance:
-	// 1. Lock, grab the last instance in the array, unlock
-	// 2. Acquire the intsance's semaphore so we can be sure anyone vying for its attention is done with it
-	// 3. Lock, shorten the array so that instance is no longer part of it, unlock
-	// 3. Kill off the internal wasm instance and then deallocate the whole thing so it gets garbage collected
-
-	// 1.
-	w.lock.RLock()
-
-	if len(w.instances) == 0 {
-		w.lock.RUnlock()
-		return nil
-	}
-
-	inst := w.instances[len(w.instances)-1]
-
-	// unlock before attempting to acquire the instance, they can deadlock eachother
-	w.lock.RUnlock()
-
-	// 2.
-	if err := inst.access.Acquire(context.Background(), 1); err != nil {
-		return errors.Wrap(err, "failed to Acquire instance for removal")
-	}
-
-	// 3.
-	w.lock.Lock()
-	if w.instIndex == len(w.instances)-1 {
-		w.instIndex--
-	}
-
-	w.instances = w.instances[:len(w.instances)-1]
-	w.lock.Unlock()
+	// grab an instance from the available queue
+	// and we won't give it back becuase it's being destroyed
+	inst := <-w.availableInstances
 
 	// 4.
 	inst.wasmerInst.Close()
@@ -200,42 +114,24 @@ func (w *wasmEnvironment) removeInstance() error {
 	inst.ffiResult = nil
 	inst.resultChan = nil
 	inst.errChan = nil
-
-	inst.access = nil
 	inst = nil
 
 	return nil
 }
 
-// useInstance provides an instance from the environment's pool to be used
+// useInstance provides an instance from the environment's pool to be used by a callback function
 func (w *wasmEnvironment) useInstance(ctx *rt.Ctx, instFunc func(*wasmInstance, int32)) error {
-	// we have to do a lock dance between w.lock and inst.access to ensure that
-	// a single instance isn't used by more than one runnable at the same time
-	w.lock.Lock()
+	// grab an instance from the available queue and then
+	// return it to the environment when finished
+	inst := <-w.availableInstances
 
-	if w.instIndex == len(w.instances)-1 {
-		w.instIndex = 0
-	} else {
-		w.instIndex++
-	}
-
-	instIndex := w.instIndex
-	inst := w.instances[instIndex]
-
-	// now that we've got an instance, release the lock since this can deadlock with inst.access
-	w.lock.Unlock()
-
-	// acquire the instance's semaphore so we are guaranteed to be the only one using it
-	acquireCtx, _ := context.WithTimeout(context.Background(), time.Duration(time.Second*10))
-	if err := inst.access.Acquire(acquireCtx, 1); err != nil {
-		return errors.Wrap(err, "failed to Acquire instance")
-	}
-
-	defer inst.access.Release(1)
+	defer func() {
+		w.availableInstances <- inst
+	}()
 
 	// generate a random identifier as a reference to the instance in use to
 	// easily allow the Wasm module to reference itself when calling back over the FFI
-	ident, err := setupNewIdentifier(w.UUID, instIndex)
+	ident, err := setupNewIdentifier(inst)
 	if err != nil {
 		return errors.Wrap(err, "failed to setupNewIdentifier")
 	}
@@ -307,162 +203,7 @@ func (w *wasmEnvironment) internals() (*wasmer.Module, *wasmer.Store, *wasmer.Im
 	return w.module, w.store, w.imports, nil
 }
 
-func setupNewIdentifier(envUUID string, instIndex int) (int32, error) {
-	for {
-		ident, err := randomIdentifier()
-		if err != nil {
-			return -1, errors.Wrap(err, "failed to randomIdentifier")
-		}
-
-		// ensure we don't accidentally overwrite something else
-		// (however unlikely that may be)
-		if _, exists := instanceMapper.Load(ident); exists {
-			continue
-		}
-
-		ref := instanceReference{
-			EnvUUID:   envUUID,
-			InstIndex: instIndex,
-		}
-
-		instanceMapper.Store(ident, ref)
-
-		return ident, nil
-	}
-}
-
-func removeIdentifier(ident int32) {
-	instanceMapper.Delete(ident)
-}
-
-func instanceForIdentifier(ident int32, needsFFIResult bool) (*wasmInstance, error) {
-	rawRef, exists := instanceMapper.Load(ident)
-	if !exists {
-		return nil, errors.New("instance does not exist")
-	}
-
-	ref := rawRef.(instanceReference)
-
-	envLock.RLock()
-	defer envLock.RUnlock()
-
-	env, exists := environments[ref.EnvUUID]
-	if !exists {
-		return nil, errors.New("environment does not exist")
-	}
-
-	inst, err := env.instanceAtIndex(ref.InstIndex)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to instanceAtIndex")
-	}
-
-	if needsFFIResult && inst.ffiResult != nil {
-		return nil, errors.New("cannot use instance for host call with existing call in progress")
-	}
-
-	return inst, nil
-}
-
-func randomIdentifier() (int32, error) {
-	// generate a random number between 0 and the largest possible int32
-	num, err := rand.Int(rand.Reader, big.NewInt(math.MaxInt32))
-	if err != nil {
-		return -1, errors.Wrap(err, "failed to rand.Int")
-	}
-
-	return int32(num.Int64()), nil
-}
-
 // UseInternalLogger sets the logger to be used log internal wasm runtime messages
 func UseInternalLogger(l *vlog.Logger) {
 	internalLogger = l
-}
-
-/////////////////////////////////////////////////////////////////////////////
-// below is the wasm glue code used to manipulate wasm instance memory     //
-// this requires a set of functions to be available within the wasm module //
-// - allocate                                                              //
-// - deallocate                                                            //
-/////////////////////////////////////////////////////////////////////////////
-
-func (w *wasmInstance) setFFIResult(data []byte) error {
-	if w.ffiResult != nil {
-		return errors.New("instance ffiResult is already set")
-	}
-
-	w.ffiResult = data
-
-	return nil
-}
-
-func (w *wasmInstance) useFFIResult() ([]byte, error) {
-	if w.ffiResult == nil {
-		return nil, errors.New("instance ffiResult is not set")
-	}
-
-	defer func() {
-		w.ffiResult = nil
-	}()
-
-	return w.ffiResult, nil
-}
-
-func (w *wasmInstance) readMemory(pointer int32, size int32) []byte {
-	memory, err := w.wasmerInst.Exports.GetMemory("memory")
-	if err != nil || memory == nil {
-		// we failed
-		return []byte{}
-	}
-
-	data := memory.Data()[pointer:]
-	result := make([]byte, size)
-
-	for index := 0; int32(index) < size; index++ {
-		result[index] = data[index]
-	}
-
-	return result
-}
-
-func (w *wasmInstance) writeMemory(data []byte) (int32, error) {
-	lengthOfInput := len(data)
-
-	allocate, err := w.wasmerInst.Exports.GetFunction("allocate")
-	if err != nil || allocate == nil {
-		return -1, errors.New("missing required FFI function: allocate")
-	}
-
-	// Allocate memory for the input, and get a pointer to it.
-	allocateResult, err := allocate(lengthOfInput)
-	if err != nil {
-		return -1, errors.Wrap(err, "failed to call allocate")
-	}
-
-	pointer := allocateResult.(int32)
-
-	w.writeMemoryAtLocation(pointer, data)
-
-	return pointer, nil
-}
-
-func (w *wasmInstance) writeMemoryAtLocation(pointer int32, data []byte) {
-	memory, err := w.wasmerInst.Exports.GetMemory("memory")
-	if err != nil || memory == nil {
-		// we failed
-		return
-	}
-
-	scopedMemory := memory.Data()[pointer:]
-
-	copy(scopedMemory, data)
-}
-
-func (w *wasmInstance) deallocate(pointer int32, length int) {
-	dealloc, err := w.wasmerInst.Exports.GetFunction("deallocate")
-	if err != nil || dealloc == nil {
-		// we failed
-		return
-	}
-
-	dealloc(pointer, length)
 }
